+++ origins/orgUnitMembership/demoOrgUnitMembership.php
@@ -29,9 +29,9 @@ class demoOrgUnitMembership extends AbstractOriginImplementation {
 	 * @return bool
 	 */
 	public function connect(): bool {
-		$csv_file = $this->config()->getFilePath();
+		$csv_file = $this->config()->getPath();
 
-		if ($this->config()->getConnectionType() != IOriginConfig::CONNECTION_TYPE_FILE || !file_exists($csv_file)) {
+		if ($this->config()->getConnectionType() != IOriginConfig::CONNECTION_TYPE_PATH || !file_exists($csv_file)) {
 			// CSV file not found!
 			throw new ConnectionFailedException("The csv file $csv_file does not exists!");
 		}
@@ -51,7 +51,7 @@ class demoOrgUnitMembership extends AbstractOriginImplementation {
 	 * @return int
 	 */
 	public function parseData(): int {// Parse csv file
-		$csv_file = $this->config()->getFilePath();
+		$csv_file = $this->config()->getPath();
 
 		$csv = new ilCSVReader();
 		$csv->setSeparator(",");