+++ src/Origin/Config/Group/GroupOriginConfig.php
@@ -28,7 +28,7 @@ class GroupOriginConfig extends OriginConfig implements IGroupOriginConfig {
 	/**
 	 * @inheritdoc
 	 */
-	public function getParentRefIdIfNoParentIdFound() {
-		return $this->data[self::REF_ID_NO_PARENT_ID_FOUND];
+	public function getParentRefIdIfNoParentIdFound(): int {
+		return intval($this->get(self::REF_ID_NO_PARENT_ID_FOUND));
 	}
 }